+++ app/src/main/java/com/example/nontonkuy/ui/movie/main/MovieListFragment.kt
@@ -83,7 +83,7 @@ class MovieListFragment : Fragment() , MovieAdapter.OnItemClickCallback{
 
     override fun onItemClicked(data: MovieEntity) {
         val intent = Intent(context, MovieDetailActivity::class.java)
-        intent.putExtra(ID_MOVIE, data.id.toInt())
+        intent.putExtra(ID_MOVIE, data.id)
         context?.startActivity(intent)
     }
