+++ app/src/main/java/com/example/nontonkuy/ui/tvshow/favorite/TvShowFavoriteFragment.kt
@@ -69,7 +69,7 @@ class TvShowFavoriteFragment : Fragment(), TvShowAdapter.OnItemClickCallback {
 
     override fun onItemClicked(data: TvShowEntity) {
         val intent = Intent(context, TvShowDetailActivity::class.java)
-        intent.putExtra(ID_TVSHOW, data.id.toInt())
+        intent.putExtra(ID_TVSHOW, data.id)
         context?.startActivity(intent)
     }
